% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oaColorFunctions.R
\name{colorWheel}
\alias{colorWheel}
\title{Generate a Color Wheel}
\usage{
colorWheel(colVec = NULL)
}
\arguments{
\item{colVec}{vector of colors}
}
\description{
Generate a Color Wheel
}
\author{
Jason Waddell
}

