% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oaColorFunctions.R
\name{oaColors}
\alias{oaColors}
\title{Pick One or More OA Colors}
\usage{
oaColors(color = NULL, alpha = 1)
}
\arguments{
\item{color}{a character vector of color names; possible values are "red", "orange", "yellow", "green", "cyan", "blue", "pink",
"limegreen", "purple", "black", "white", "grey" or "gray"}

\item{alpha}{transparency level for the color(s)}
}
\value{
character vector of colors
}
\description{
Pick One or More OA Colors
}
\author{
Tobias Verbeke
}

